requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.2.0
langchain==0.3.0
langchain-huggingface==0.1.0
langchain-community==0.3.0
pypdf==5.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
